feat: add components getMenu, RenderContent. Modify css and include home.tsx
