+++ frontend/src/pages/Home/style.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Press+Start+2P&family=Roboto&family=Roboto+Condensed:wght@300&display=swap');
+
 .header {
   background: #001529;
   top: 0;
@@ -12,4 +14,81 @@
 
 .breadcrumb {
   margin: 16px 0;
+}
+
+.divider {
+  border-color: #001529;
+  margin-bottom: 60px;
+}
+
+.text {
+  color: #001529;
+  text-align: center;
+  font-family: Roboto Condensed;
+  font-size: 3rem;
+}
+
+.pong-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+  height: 500px;
+  max-width: 60%;
+  border: 2px solid #001529;
+  border-radius: 9px;
+  box-shadow: 0 0 20px #001529;
+  margin: 0 auto;
+}
+
+.middle-line {
+  position: absolute;
+  left: 50%;
+  top: 0;
+  bottom: 0;
+  width: 2px;
+  background-color: #001529;
+  transform: translateX(-50%);
+}
+
+@keyframes moveUp {
+  0%,
+  100% {
+    top: 400px;
+  }
+  50% {
+    top: 0;
+  }
+}
+
+@keyframes moveDown {
+  0%,
+  100% {
+    top: 0;
+  }
+  50% {
+    top: 400px;
+  }
+}
+
+.paddle {
+  width: 20px;
+  height: 100px;
+  background-color: #001529;
+  position: absolute;
+  top: 0;
+  animation-duration: 3s;
+  animation-iteration-count: infinite;
+  animation-timing-function: linear;
+}
+
+.paddle-left {
+  left: 0;
+  animation-name: moveDown;
+}
+
+.paddle-right {
+  right: 0;
+  animation-name: moveUp;
 }
